Added delete dialog

+++ src/app/home/home.component.ts
@@ -101,23 +101,35 @@ export class HomeComponent implements OnInit {
 
   //TODO: Add dialog box for deleting card
   deleteCard(selectedCard: Card): void { //TODO: Need to handle errors
-    this.apiService.patchDeckRemoveCard(this.deck.id, selectedCard.id).toPromise().then(deck => {
-      console.log("Card deleted from deck");
-      this.deck = deck;
-      this.cardSearch(this.cardSearchTerm);
-      this.apiService.getDecks().toPromise().then((decks) => {this.allDecks = decks['decks']});
+    const dialogRef = this.dialog.open(DeleteDialog);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+          this.apiService.patchDeckRemoveCard(this.deck.id, selectedCard.id).toPromise().then(deck => {
+          console.log("Card deleted from deck");
+          this.deck = deck;
+          this.cardSearch(this.cardSearchTerm);
+          this.apiService.getDecks().toPromise().then((decks) => {this.allDecks = decks['decks']});
+        });
+      }   
     });
   }
 
 
   //TODO: Add dialog box for deleting deck
   deleteDeck(selectedDeck: Deck): void {
-    this.apiService.deleteDeck(selectedDeck.id).toPromise().finally(() => {
-      this.apiService.getDecks().toPromise().then(decks => {
-        this.allDecks = decks['decks'];
-        this.searchedDecks = this.allDecks;
-      });
-    }) //TODO: maybe do something with the data idk.
+    const dialogRef = this.dialog.open(DeleteDialog);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+            this.apiService.deleteDeck(selectedDeck.id).toPromise().finally(() => {
+          this.apiService.getDecks().toPromise().then(decks => {
+            this.allDecks = decks['decks'];
+            this.searchedDecks = this.allDecks;
+          });
+        }) //TODO: maybe do something with the data idk.
+      }
+    });
   }
 
   createNewCard(newCard: CardPutt): void {
@@ -224,6 +236,18 @@ export class EditCardDialog {
       public dialogRef: MatDialogRef<EditCardDialog>,
       @Inject(MAT_DIALOG_DATA) public data: Card
     ) {}
+
+  onNoClick(): void {
+    this.dialogRef.close();
+  }
+}
+
+@Component({
+  selector: 'delete-dialog',
+  templateUrl: 'delete-dialog.html',
+})
+export class DeleteDialog {
+  constructor(public dialogRef: MatDialogRef<DeleteDialog>) {}
 
   onNoClick(): void {
     this.dialogRef.close();
